fix: merge resource mutation tests to satisfy pkg-file-count
Fold accepted mixed work/resource mutation coverage into subsystem_transitioner_failure_resource_test.go so pkg/factory/subsystems stays within the 15-file package limit.
